Update ServerTastic plugin to work with apiv2
- Added new Positive SSL types to mapped type ids. These new types
were added to the constants in SSLPlugin.php.
- Renamed Symantec type ids to Verisign. New names were also updated
for the constants in SSLPlugin.php
- Reorderd the mapped type ids for better readability
- Updated plugin actions variables and getPluginActions variables
to comply with API version 2 functionality
- Added PHP doc blocks for all methods
- Removed 4 year product option as ServerTastic does not offer this
for any of their products anymore
- Added a method to decode the CSR from the UI, as we need the domain
name from it in order to properly get the approved emails from the API.
This uses the openssl php method to decode it and will throw an exception
if the extension is not installed/loaded. ServerTastic will eventually
have an API endpoint to do this so a todo was added.
- Removed the invite email methods as they no longer apply with API
version 2.
- Added a POST option to the make request method as it is required by
the new API when sending the CSR as an argument.